Redesign of tckedit and tckresample
Command tckresample is now exclusively responsible for operations that involve re-sampling streamlines trajectories.
In addition to the line- and arc-based path resampling that were previously available, it now also performs other types of resampling; some of which were previously found in tckedit, as well as a new resampling feature (having a fixed number of points per streamline).
The resampling functionality has been removed from tckedit, as it was awkward to maintain and determine the appropriate order of operations when combined with other functionalities in that command.
A lot of the streamline resampling code has now been moved to a dedicated directory, src/dwi/tractography/resampling/.
