+++ src/dwi/tractography/editing/editing.cpp
@@ -42,26 +42,6 @@ const OptionGroup LengthOption = OptionGroup ("Streamline length threshold optio
 
 
 
-
-const OptionGroup ResampleOption = OptionGroup ("Streamline resampling options")
-
-  + Option ("upsample", "increase the density of points along the length of the streamline by some factor "
-                        "(may improve mapping streamlines to ROIs, and/or visualisation)")
-    + Argument ("ratio").type_integer (1, 1, 1e6)
-
-  + Option ("downsample", "increase the density of points along the length of the streamline by some factor "
-                          "(decreases required storage space)")
-    + Argument ("ratio").type_integer (1, 1, 1e6)
-
-  + Option ("resample", "re-sample the streamline to a desired step size (in mm)")
-    + Argument ("step_size").type_float (0.0, 0.0, 1e6)
-
-  + Option ("out_ends_only", "only output the two endpoints of each streamline");
-
-
-
-
-
 const OptionGroup TruncateOption = OptionGroup ("Streamline count truncation options")
 
   + Option ("number", "set the desired number of selected streamlines to be propagated to the output file")
@@ -135,10 +115,6 @@ void load_properties (Tractography::Properties& properties)
       }
     }
   }
-
-
-  // ResampleOption
-  // The relevant entry in Properties is updated at a later stage
 
 
   // TruncateOption
