+++ src/dwi/tractography/editing/receiver.h
@@ -40,11 +40,10 @@ namespace MR {
 
           public:
 
-            Receiver (const std::string& path, const Properties& properties, const size_t n, const size_t s, const bool e) :
+            Receiver (const std::string& path, const Properties& properties, const size_t n, const size_t s) :
               writer (path, properties),
               number (n),
               skip (s),
-              ends_only (e),
               // Need to use local counts instead of writer class members due to track cropping
               count (0),
               total_count (0),
@@ -67,11 +66,8 @@ namespace MR {
             Writer<> writer;
             const uint64_t number;
             uint64_t skip;
-            const bool ends_only;
             uint64_t count, total_count;
             ProgressBar progress;
-
-            void output (const Streamline<>&);
 
         };
